Adding required styles
Our telemetry system can't target empty divs, but if you add a minHeight, it works. Not sure why, but I think this is a benign style that will work for us.

The marginTop is to match the h3 above, in experimenting we need some sort of margin, I'm open to changes here.