Use awareness store for debug data user mapping
Moves Yjs client-to-user mapping from YDoc state persistence to
awareness store's in-memory tracking. We had issues with getting that
to sync correctly. This is simpler to start with than looking into why
it's not working. This has the trade-off of only including users seen
after current client joined, but this is acceptable as first iteration.